fix: Add conftest.py and clients.py files in core testing packages
