+++ tests/pickle_tests/conftest.py
@@ -1,0 +1,1 @@
+"""Shared fixtures for pickle core tests."""
